moved fixtures that are used by tabular tests to "tests/tabular/conftest.py"
