+++ tests/tabular/test_datasets.py
@@ -12,11 +12,12 @@ import sys
 
 import numpy as np
 import sklearn
-from deepchecks.tabular.datasets.classification import adult, breast_cancer, iris, lending_club, phishing
-from deepchecks.tabular.datasets.regression import avocado, wine_quality
 from deepdiff import DeepDiff
 from hamcrest import assert_that, instance_of
 from sklearn.base import BaseEstimator
+
+from deepchecks.tabular.datasets.classification import adult, breast_cancer, iris, lending_club, phishing
+from deepchecks.tabular.datasets.regression import avocado, wine_quality
 
 
 def assert_sklearn_model_params_equals(model1, model2):
