+++ tests/utils/preprocessing_test.py
@@ -13,8 +13,8 @@ import numpy as np
 from hamcrest import assert_that, calling, equal_to, raises
 
 from deepchecks.core.errors import DeepchecksValueError
-from deepchecks.utils.distribution.preprocessing import (OTHER_CATEGORY_NAME,
-                                                         preprocess_2_cat_cols_to_same_bins, value_frequency)
+from deepchecks.utils.distribution.preprocessing import (OTHER_CATEGORY_NAME, preprocess_2_cat_cols_to_same_bins,
+                                                         value_frequency)
 
 
 def test_cat_cols_to_bins_no_max_num_categories():
