+++ tests/vision/base/test_vision_data.py
@@ -13,17 +13,16 @@
 import numpy as np
 import pytest
 import torch
-from hamcrest import assert_that, calling, equal_to, raises, is_not, has_length
+from hamcrest import assert_that, calling, equal_to, has_length, is_not, raises
 from torch.utils.data import DataLoader
 
-from deepchecks.core.errors import ValidationError, DatasetValidationError
-from deepchecks.vision.datasets.classification.mnist import collate_without_model, IterableTorchMnistDataset
+from deepchecks.core.errors import DatasetValidationError, ValidationError
+from deepchecks.vision.datasets.classification.mnist import IterableTorchMnistDataset, collate_without_model
 from deepchecks.vision.datasets.detection import coco_torch
 from deepchecks.vision.datasets.segmentation import segmentation_coco
-from deepchecks.vision.vision_data import TaskType
 from deepchecks.vision.utils.test_utils import replace_collate_fn_dataloader
+from deepchecks.vision.vision_data import TaskType, VisionData
 from deepchecks.vision.vision_data.utils import validate_vision_data_compatibility
-from deepchecks.vision.vision_data import VisionData
 from tests.vision.conftest import run_update_loop
